music_api_django/requirements.txt
django
djangorestframework
pymysql
python-dotenv
requests
